# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ExecFleet API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/driver/attendance": {
      "post": {
        "tags": [
          "Drivers"
        ],
        "summary": "Mark driver attendance",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "driver_id": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  }
                },
                "required": [
                  "driver_id",
                  "date"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Attendance marked",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/driver/seed-schedules": {
      "post": {
        "tags": [
          "Drivers",
          "Schedule"
        ],
        "summary": "Seed dummy schedules for drivers",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "driver_id": {
                      "type": "string"
                    },
                    "date": {
                      "type": "string",
                      "format": "date"
                    },
                    "stops": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "office": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "driver_id",
                    "date",
                    "stops",
                    "office"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Schedules seeded"
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/admin/vehicle": {
      "post": {
        "summary": "Create a new vehicle with image upload",
        "tags": [
          "Admin",
          "Vehicle"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Array of image files"
                  },
                  "first_name": {
                    "type": "string"
                  },
                  "last_name": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "registration_city": {
                    "type": "string"
                  },
                  "registration_number": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "lat": {
                    "type": "number",
                    "format": "float"
                  },
                  "lng": {
                    "type": "number",
                    "format": "float"
                  },
                  "driver_id": {
                    "type": "string",
                    "description": "MongoDB ObjectId of the driver"
                  }
                },
                "required": [
                  "images",
                  "first_name",
                  "last_name",
                  "phone",
                  "type",
                  "registration_city",
                  "registration_number",
                  "address",
                  "lat",
                  "lng",
                  "driver_id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vehicle created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Vehicle ID"
                    },
                    "first_name": {
                      "type": "string"
                    },
                    "last_name": {
                      "type": "string"
                    },
                    "phone": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "registration_city": {
                      "type": "string"
                    },
                    "registration_number": {
                      "type": "string"
                    },
                    "address": {
                      "type": "string"
                    },
                    "driver_name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Vehicle with this registration number already exists"
          },
          "404": {
            "description": "Driver not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/admin/vehicle": {
      "get": {
        "tags": [
          "Admin",
          "Vehicle"
        ],
        "summary": "Get vehicle by ID",
        "description": "Fetches a vehicle's details using its `vehicle_id`.",
        "parameters": [
          {
            "name": "vehicle_id",
            "in": "query",
            "required": true,
            "description": "MongoDB ObjectId of the vehicle",
            "schema": {
              "type": "string",
              "example": "64cabc1234567890abcdef12"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "64cabc1234567890abcdef12"
                    },
                    "image_urls": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "format": "uri"
                      }
                    },
                    "first_name": {
                      "type": "string",
                      "example": "Ali"
                    },
                    "last_name": {
                      "type": "string",
                      "example": "Raza"
                    },
                    "phone": {
                      "type": "string",
                      "example": "03001234567"
                    },
                    "type": {
                      "type": "string",
                      "example": "Car"
                    },
                    "registration_city": {
                      "type": "string",
                      "example": "Karachi"
                    },
                    "registration_number": {
                      "type": "string",
                      "example": "ABC-123"
                    },
                    "address": {
                      "type": "string",
                      "example": "Street 123"
                    },
                    "location": {
                      "type": "object",
                      "properties": {
                        "lat": {
                          "type": "number",
                          "example": 24.8607
                        },
                        "lng": {
                          "type": "number",
                          "example": 67.0011
                        }
                      }
                    },
                    "driver_id": {
                      "type": "string",
                      "example": "64df1234567890abcdef12"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid vehicle_id"
          },
          "404": {
            "description": "Vehicle not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}